fix: 修复编码异常

+++ src/main/java/com/anyilanxin/msgpack/value/DoubleValue.java
@@ -57,7 +57,7 @@ public class DoubleValue extends BaseValue {
 
   @Override
   public void read(final MsgPackReader reader) {
-    value = reader.readFloat();
+    value = reader.readDouble();
   }
 
   @Override
